feat(ledger): widen cluster_wealth_db accumulator u64 -> u128 (#626 PR 2/3) (#628)
Widen the per-cluster wealth accumulator from u64 to u128 so cumulative
tagged wealth can exceed the former u64::MAX pico ceiling (18.44M BTH),
which the log-domain factor curve (#626 PR 1) now maps across the full
supply range. Consensus-breaking; rides the protocol-4.0.0 fresh-genesis
reset (bundled per the #606 precedent).

Storage / determinism (botho/src/ledger/store.rs):
- Serialization contract pinned: cluster_wealth_db values are 16-byte LE
  u128, decoded via a single `decode_cluster_wealth` helper.
- Reject-legacy posture (fail closed, M7 discipline): a wrong-width (e.g.
  8-byte) value is a hard LedgerError, never silently reinterpreted. Fresh
  genesis is assumed; there is no in-place migration.
- Incremental (`update_cluster_wealth_for_output`) and rebuild
  (`rebuild_cluster_wealth_index`) paths keep full u128 math with
  saturating_add — byte-identical semantics preserved at the new width
  (the M3/#604/#607 parity contract).
- Readers return u128: `get_cluster_wealth`, `get_all_cluster_wealth`,
  `compute_cluster_wealth_for_utxos` / `ClusterWealthInfo`, the consensus
  lottery-tilt path, and the snapshot create/load path (UtxoSnapshot
  cluster wealth widened to u128).

Consumer seam (u128 accumulator, u64 fee/centroid APIs until PR 3):
- Call sites feeding the still-u64 APIs (mempool relay fee, ring-centroid
  fee floor, CLI send estimate) clamp to u64 deterministically; the
  lottery tilt uses the un-clamped u128 directly (factor() accepts u128).

RPC (u128 exceeds JS number range -> string-encoded, per totalMined
precedent): cluster_getWealth, cluster_getWealthByTargetKeys,
cluster_getAllWealth, estimateFee and entropy_estimateFee now emit wealth
fields as decimal strings and accept the cluster_wealth param as a string
(numeric still accepted). Web adapter already sends/reads these as
bigint/string; no web or mobile parser change required.

Tests: rebuild-vs-incremental parity summing PAST u64::MAX (exact, no
saturation); 16-byte LE serialization pin; legacy 8-byte rejection on
every reader; u128-boundary saturating_add discipline.

Updates #626.

+++ botho/src/ledger/snapshot.rs
@@ -111,7 +111,7 @@ impl UtxoSnapshot {
         chain_state: ChainState,
         utxos: Vec<Utxo>,
         key_images: Vec<([u8; 32], u64)>, // (key_image, spent_height)
-        cluster_wealth: Vec<(u64, u64)>,  // (cluster_id, wealth)
+        cluster_wealth: Vec<(u64, u128)>, // (cluster_id, wealth) — u128 pico (#626)
     ) -> Result<Self, SnapshotError> {
         // Compute UTXO Merkle root
         let utxo_hashes: Vec<[u8; 32]> = utxos
@@ -178,7 +178,7 @@ impl UtxoSnapshot {
     }
 
     /// Decompress and deserialize the cluster wealth data.
-    pub fn get_cluster_wealth(&self) -> Result<Vec<(u64, u64)>, SnapshotError> {
+    pub fn get_cluster_wealth(&self) -> Result<Vec<(u64, u128)>, SnapshotError> {
         let decompressed = decompress(&self.cluster_wealth_data)?;
         bincode::deserialize(&decompressed).map_err(|e| SnapshotError::Serialization(e.to_string()))
     }
@@ -444,7 +444,7 @@ mod tests {
     fn test_snapshot_creation() {
         let utxos = vec![test_utxo(1), test_utxo(2), test_utxo(3)];
         let key_images = vec![([1u8; 32], 100u64), ([2u8; 32], 200u64)];
-        let cluster_wealth = vec![(1u64, 1000u64), (2u64, 2000u64)];
+        let cluster_wealth = vec![(1u64, 1000u128), (2u64, 2000u128)];
 
         let snapshot = UtxoSnapshot::new(
             1000,
